+++ packages/froala/attribute.js
@@ -13,4 +13,3 @@ orion.attributes.registerAttribute('froala', {
 
 Options.init('froala.height');
 orion.config.add('FROALA_ACTIVATION_KEY', 'froala', { public: true });
-