+++ packages/relationships/package.js
@@ -1,7 +1,7 @@
 Package.describe({
   name: 'orionjs:relationships',
   summary: 'Define and use relationships between meteor collections, entities and the dictionary',
-  version: '1.6.0',
+  version: '1.7.0',
   git: 'https://github.com/orionjs/orion'
 });
 
@@ -9,14 +9,16 @@ Package.onUse(function(api) {
   api.versionsFrom('1.0');
 
   api.use([
-    'orionjs:base@1.6.0',
-    'orionjs:attributes@1.6.0',
+    'orionjs:base@1.7.0',
+    'orionjs:attributes@1.7.0',
     'less@2.5.0_1',
-    'jeremy:selectize@0.12.1'
+    'jeremy:selectize@0.12.1',
+    'reactive-var'
     ]);
 
   api.imply([
     'jeremy:selectize',
+    'reactive-var'
     ]);
 
   api.addFiles([